requirements.txt
Django==1.8.1
gunicorn==19.3.0
psycopg2==2.6
redis==2.10.3
celery
django-celery
django-hstore==1.3.5
djangorestframework==3.1.2
dj-database-url
django-filter
raven==5.1.1
dj-static
